Move 'todo_lists' and 'todos' apps into a new folder 'labdale' inside outermost 'labdale'
