+++ labdale/todo_lists/apps.py
@@ -2,4 +2,4 @@ from django.apps import AppConfig
 
 
 class TodoListsConfig(AppConfig):
-    name = "todo_lists"
+    name = "labdale.todo_lists"
