+++ labdale/todos/apps.py
@@ -2,4 +2,4 @@ from django.apps import AppConfig
 
 
 class TodosConfig(AppConfig):
-    name = 'todos'
+    name = 'labdale.todos'
