+++ labdale/todos/models.py
@@ -1,6 +1,6 @@
 from django.db import models
 
-from todo_lists.models import TodoList
+from labdale.todo_lists.models import TodoList
 
 class Todo(models.Model):
     title = models.CharField(max_length=255)
